# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Sun Ready project management system",
        "title": "Sun Ready API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@sunready.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/companies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all companies with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "List all companies",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by referred by user ID",
                        "name": "referred_by_user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompaniesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new company with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Create a new company",
                "parameters": [
                    {
                        "description": "Company details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/companies/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all companies without pagination with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Get all companies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by referred by user ID",
                        "name": "referred_by_user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompaniesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/companies/slug/{slug}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a company by its slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Get company by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a company by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Get company by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a company's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Update company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Company update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a company by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Delete company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CompaniesResponse": {
            "type": "object",
            "properties": {
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Company"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handler.CompanyResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/models.Company"
                }
            }
        },
        "handler.CreateCompanyRequest": {
            "type": "object",
            "properties": {
                "baseline_adder": {
                    "type": "number",
                    "example": 100
                },
                "code": {
                    "type": "string",
                    "example": "ACME"
                },
                "description": {
                    "type": "string",
                    "example": "Leading solar company"
                },
                "display_name": {
                    "type": "string",
                    "example": "Acme Corporation"
                },
                "logo_path": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Acme Corp"
                },
                "sales_commission_default": {
                    "type": "number",
                    "example": 0.1
                },
                "sales_commission_max": {
                    "type": "number",
                    "example": 0.15
                },
                "sales_commission_min": {
                    "type": "number",
                    "example": 0.05
                },
                "slug": {
                    "type": "string",
                    "example": "acme-corp"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "handler.UpdateCompanyRequest": {
            "type": "object",
            "properties": {
                "baseline": {
                    "type": "number",
                    "example": 1000
                },
                "baseline_adder": {
                    "type": "number",
                    "example": 100
                },
                "baseline_adder_pct_sales_comms": {
                    "type": "integer",
                    "example": 10
                },
                "code": {
                    "type": "string",
                    "example": "ACME"
                },
                "contract_tag": {
                    "type": "string",
                    "example": "STANDARD"
                },
                "description": {
                    "type": "string",
                    "example": "Leading solar company"
                },
                "display_name": {
                    "type": "string",
                    "example": "Acme Corporation"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "logo_path": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Acme Corp"
                },
                "sales_commission_default": {
                    "type": "number",
                    "example": 0.1
                },
                "sales_commission_max": {
                    "type": "number",
                    "example": 0.15
                },
                "sales_commission_min": {
                    "type": "number",
                    "example": 0.05
                },
                "slug": {
                    "type": "string",
                    "example": "acme-corp"
                }
            }
        },
...
        "models.Company": {
            "type": "object",
            "properties": {
                "admin_id": {
                    "type": "integer",
                    "example": 1
                },
                "baseline": {
                    "type": "number",
                    "example": 1000
                },
                "baseline_adder": {
                    "type": "number",
                    "example": 100
                },
                "baseline_adder_pct_sales_comms": {
                    "type": "integer",
                    "example": 10
                },
                "code": {
                    "type": "string",
                    "example": "ACME"
                },
                "contract_tag": {
                    "type": "string",
                    "example": "STANDARD"
                },
                "created_at": {
                    "type": "string"
                },
                "credits": {
                    "type": "integer",
                    "example": 1000
                },
                "custom_commissions": {
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "type": "string",
                    "example": "Leading solar company"
                },
                "display_name": {
                    "type": "string",
                    "example": "Acme Corporation"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "logo_path": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Acme Corp"
                },
                "pricing_mode": {
                    "type": "integer",
                    "example": 0
                },
                "referred_by_user_id": {
                    "type": "integer",
                    "example": 1
                },
                "sales_commission_default": {
                    "type": "number",
                    "example": 0.1
                },
                "sales_commission_max": {
                    "type": "number",
                    "example": 0.15
                },
                "sales_commission_min": {
                    "type": "number",
                    "example": 0.05
                },
                "slug": {
                    "type": "string",
                    "example": "acme-corp"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}